Myportfolio_Project/backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
django-environ==0.12.0
djangorestframework==3.16.1
dotenv==0.9.9
feedparser==6.0.12
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
sentry-sdk==2.38.0
sgmllib3k==1.0.0
sqlparse==0.5.3
tweepy==4.16.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0
